+++ internal/constants.go
@@ -3,4 +3,8 @@ package internal
 const (
 	ConvertPNG string = "Convert To PNG"
 	ConvertPDF string = "Convert To PDF"
+
+	DeviceAuto  string = "Infer by note"
+	DeviceManta string = "Supernote Manta X2"
+	DeviceNomad string = "Supernote Nomad X2"
 )