+++ internal/header.go
@@ -7,7 +7,6 @@ import (
 type Header struct {
 	APPLY_EQUIPMENT  string
 	FILE_RECOGN_TYPE string
-	HORIZONTAL_CHECK string
 }
 
 func NewHeader(file *os.File, notebook *Notebook) error {
@@ -21,7 +20,6 @@ func NewHeader(file *os.File, notebook *Notebook) error {
 	notebook.Header = &Header{
 		APPLY_EQUIPMENT:  metadata["APPLY_EQUIPMENT"],
 		FILE_RECOGN_TYPE: metadata["FILE_RECOGN_TYPE"],
-		HORIZONTAL_CHECK: metadata["HORIZONTAL_CHECK"],
 	}
 
 	return nil